Eliminated all warnings in the compilation process, except for the ones related to Getline.c which should be solved differently.
git-svn-id: https://svn.cern.ch/reps/psi46expert/trunk@125 4525493e-7705-40b1-a816-d608a930855b

+++ src/psi46expert/ThrComp.cc
@@ -42,7 +42,7 @@ void ThrComp::RocAction()
   SetDAC("CtrlReg", 0);
 
   TGraph* graph = new TGraph();
-  TString name = Form("VthrComp_Vcal%i", vcal);
+  TString name = Form("VthrComp_Vcal%i", (int) vcal);
   graph->SetName(name);
   graph->SetLineColor(2);
   graph->SetLineStyle(1);
